# Speed up boolean comparison kernels (~3x)