* Reuse `binary` and `unary` functions to do computation on `u64` elements and cleanup code.
* Restore comparison benchmark

Outcome:

```
bool 2^10               time:   [311.72 ns 312.34 ns 313.03 ns]
                        change: [-50.796% -50.519% -50.257%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 5 outliers among 100 measurements (5.00%)
  4 (4.00%) high mild
  1 (1.00%) high severe

Benchmarking bool scalar 2^10: Collecting 100 samples in estimated 5.0002 s (36M                                                                                bool scalar 2^10        time:   [134.00 ns 134.44 ns 134.93 ns]
                        change: [-29.129% -28.672% -28.185%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 5 outliers among 100 measurements (5.00%)
  4 (4.00%) high mild
  1 (1.00%) high severe

Benchmarking bool 2^12: Collecting 100 samples in estimated 5.0011 s (7.5M itera                                                                                bool 2^12               time:   [682.72 ns 690.96 ns 698.70 ns]
                        change: [-67.207% -66.954% -66.710%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 3 outliers among 100 measurements (3.00%)
  1 (1.00%) high mild
  2 (2.00%) high severe

Benchmarking bool scalar 2^12: Collecting 100 samples in estimated 5.0010 s (25M                                                                                bool scalar 2^12        time:   [194.41 ns 197.01 ns 200.69 ns]
                        change: [-51.596% -51.032% -50.249%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 8 outliers among 100 measurements (8.00%)
  1 (1.00%) high mild
  7 (7.00%) high severe

Benchmarking bool 2^14: Collecting 100 samples in estimated 5.0098 s (2.2M itera                                                                                bool 2^14               time:   [2.2481 us 2.2573 us 2.2680 us]
                        change: [-71.747% -71.515% -71.308%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 6 outliers among 100 measurements (6.00%)
  1 (1.00%) low mild
  5 (5.00%) high mild

Benchmarking bool scalar 2^14: Collecting 100 samples in estimated 5.0003 s (11M                                                                                bool scalar 2^14        time:   [449.36 ns 450.69 ns 452.27 ns]
                        change: [-66.059% -65.853% -65.645%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 4 outliers among 100 measurements (4.00%)
  1 (1.00%) high mild
  3 (3.00%) high severe

Benchmarking bool 2^16: Collecting 100 samples in estimated 5.0003 s (616k itera                                                                                bool 2^16               time:   [8.1296 us 8.1497 us 8.1684 us]
                        change: [-75.072% -74.346% -73.690%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 6 outliers among 100 measurements (6.00%)
  1 (1.00%) low mild
  4 (4.00%) high mild
  1 (1.00%) high severe

Benchmarking bool scalar 2^16: Collecting 100 samples in estimated 5.0012 s (3.8                                                                                bool scalar 2^16        time:   [1.2945 us 1.2995 us 1.3050 us]
                        change: [-72.611% -72.412% -72.226%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 8 outliers among 100 measurements (8.00%)
  5 (5.00%) high mild
  3 (3.00%) high severe
```